+++ sdks/java/io/parquet/src/test/java/org/apache/beam/sdk/io/parquet/ParquetIOTest.java
@@ -208,27 +208,6 @@ public class ParquetIOTest implements Serializable {
   }
 
   @Test
-  public void testWriteAndReadWithSplit() {
-    List<GenericRecord> records = generateGenericRecords(1000);
-
-    mainPipeline
-        .apply(Create.of(records).withCoder(AvroCoder.of(SCHEMA)))
-        .apply(
-            FileIO.<GenericRecord>write()
-                .via(ParquetIO.sink(SCHEMA))
-                .to(temporaryFolder.getRoot().getAbsolutePath()));
-    mainPipeline.run().waitUntilFinish();
-
-    PCollection<GenericRecord> readBackWithSplit =
-        readPipeline.apply(
-            ParquetIO.read(SCHEMA)
-                .from(temporaryFolder.getRoot().getAbsolutePath() + "/*")
-                .withSplit());
-    PAssert.that(readBackWithSplit).containsInAnyOrder(records);
-    readPipeline.run().waitUntilFinish();
-  }
-
-  @Test
   public void testWriteAndReadWithBeamSchema() {
     List<GenericRecord> records = generateGenericRecords(1000);
 
@@ -252,7 +231,7 @@ public class ParquetIOTest implements Serializable {
   }
 
   @Test
-  public void testWriteAndReadFilesAsJsonForWithSplitForUnknownSchema() {
+  public void testWriteAndReadFilesAsJsonForUnknownSchema() {
     List<GenericRecord> records = generateGenericRecords(1000);
 
     mainPipeline
@@ -263,13 +242,12 @@ public class ParquetIOTest implements Serializable {
                 .to(temporaryFolder.getRoot().getAbsolutePath()));
     mainPipeline.run().waitUntilFinish();
 
-    PCollection<String> readBackAsJsonWithSplit =
+    PCollection<String> readBackAsJson =
         readPipeline.apply(
             ParquetIO.parseGenericRecords(ParseGenericRecordAsJsonFn.create())
-                .from(temporaryFolder.getRoot().getAbsolutePath() + "/*")
-                .withSplit());
+                .from(temporaryFolder.getRoot().getAbsolutePath() + "/*"));
 
-    PAssert.that(readBackAsJsonWithSplit).containsInAnyOrder(convertRecordsToJson(records));
+    PAssert.that(readBackAsJson).containsInAnyOrder(convertRecordsToJson(records));
     readPipeline.run().waitUntilFinish();
   }
 
@@ -392,7 +370,6 @@ public class ParquetIOTest implements Serializable {
         DisplayData.from(
             ParquetIO.read(SCHEMA)
                 .from("foo.parquet")
-                .withSplit()
                 .withProjection(REQUESTED_SCHEMA, SCHEMA)
                 .withAvroDataModel(GenericData.get())
                 .withConfiguration(configuration));
@@ -400,7 +377,6 @@ public class ParquetIOTest implements Serializable {
     assertThat(displayData, hasDisplayItem("filePattern", "foo.parquet"));
     assertThat(displayData, hasDisplayItem("schema", SCHEMA.toString()));
     assertThat(displayData, hasDisplayItem("inferBeamSchema", false));
-    assertThat(displayData, hasDisplayItem("splittable", true));
     assertThat(displayData, hasDisplayItem("projectionSchema", REQUESTED_SCHEMA.toString()));
     assertThat(displayData, hasDisplayItem("avroDataModel", GenericData.get().toString()));
     assertThat(displayData, hasDisplayItem("parquet.foo", "foo"));
@@ -436,29 +412,6 @@ public class ParquetIOTest implements Serializable {
     readPipeline.run().waitUntilFinish();
   }
 
-  @Test(expected = org.apache.beam.sdk.Pipeline.PipelineExecutionException.class)
-  public void testWriteAndReadWithSplitUsingReflectDataSchemaWithoutDataModelThrowsException() {
-    Schema testRecordSchema = ReflectData.get().getSchema(TestRecord.class);
-
-    List<GenericRecord> records = generateGenericRecords(1000);
-    mainPipeline
-        .apply(Create.of(records).withCoder(AvroCoder.of(testRecordSchema)))
-        .apply(
-            FileIO.<GenericRecord>write()
-                .via(ParquetIO.sink(testRecordSchema))
-                .to(temporaryFolder.getRoot().getAbsolutePath()));
-    mainPipeline.run().waitUntilFinish();
-
-    PCollection<GenericRecord> readBack =
-        readPipeline.apply(
-            ParquetIO.read(testRecordSchema)
-                .withSplit()
-                .from(temporaryFolder.getRoot().getAbsolutePath() + "/*"));
-
-    PAssert.that(readBack).containsInAnyOrder(records);
-    readPipeline.run().waitUntilFinish();
-  }
-
   @Test
   public void testWriteAndReadUsingReflectDataSchemaWithDataModel() {
     Schema testRecordSchema = ReflectData.get().getSchema(TestRecord.class);
@@ -475,30 +428,6 @@ public class ParquetIOTest implements Serializable {
     PCollection<GenericRecord> readBack =
         readPipeline.apply(
             ParquetIO.read(testRecordSchema)
-                .withAvroDataModel(GenericData.get())
-                .from(temporaryFolder.getRoot().getAbsolutePath() + "/*"));
-
-    PAssert.that(readBack).containsInAnyOrder(records);
-    readPipeline.run().waitUntilFinish();
-  }
-
-  @Test
-  public void testWriteAndReadwithSplitUsingReflectDataSchemaWithDataModel() {
-    Schema testRecordSchema = ReflectData.get().getSchema(TestRecord.class);
-
-    List<GenericRecord> records = generateGenericRecords(1000);
-    mainPipeline
-        .apply(Create.of(records).withCoder(AvroCoder.of(testRecordSchema)))
-        .apply(
-            FileIO.<GenericRecord>write()
-                .via(ParquetIO.sink(testRecordSchema))
-                .to(temporaryFolder.getRoot().getAbsolutePath()));
-    mainPipeline.run().waitUntilFinish();
-
-    PCollection<GenericRecord> readBack =
-        readPipeline.apply(
-            ParquetIO.read(testRecordSchema)
-                .withSplit()
                 .withAvroDataModel(GenericData.get())
                 .from(temporaryFolder.getRoot().getAbsolutePath() + "/*"));